move : vo패키지 모델 패키지 내부로 이동
