+++ src/main/java/christmas/model/Discount.java
@@ -10,8 +10,8 @@ import static christmas.model.DiscountPolicy.SPECIAL_DISCOUNT;
 import static christmas.model.DiscountPolicy.WEEKDAY_DISCOUNT;
 import static christmas.model.DiscountPolicy.WEEKEND_DISCOUNT;
 
-import christmas.vo.DiscountAmount;
-import christmas.vo.VisitDay;
+import christmas.model.vo.DiscountAmount;
+import christmas.model.vo.VisitDay;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
