+++ src/main/java/christmas/util/Converter.java
@@ -3,8 +3,8 @@ package christmas.util;
 import christmas.exception.AlreadyExistsInOrderException;
 import christmas.exception.InvalidOrderFormatException;
 import christmas.exception.OrderMenuCountNonNumericException;
-import christmas.vo.Food;
-import christmas.vo.Quantity;
+import christmas.model.vo.Food;
+import christmas.model.vo.Quantity;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
